chore: move interface, rename file
