--- src/translators/LLMTranslators/LLMTranslatorTypes.ts
@@ -1,26 +1,0 @@
-import { LLMTranslatorOptions } from './LLMTranslator';
-
-export interface LLMFetcher {
-	/**
-	 * Method for request to AI model
-	 */
-	fetch(prompt: string): Promise<string>;
-
-	/**
-	 * Max length of string for prompt
-	 */
-	getLengthLimit(): number;
-
-	/**
-	 * Delay between requests to comply with the requests per minute limit
-	 */
-	getRequestsTimeout(): number;
-}
-
-export type TranslatorConfig = {
-	apiKey: string;
-	apiHost: string;
-	userAgent?: string;
-	model?: string;
-	translatorOptions: LLMTranslatorOptions;
-};
